requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Deprecated==1.2.14
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gcloud==0.18.3
googleapis-common-protos==1.59.1
httplib2==0.22.0
idna==3.4
jwcrypto==1.5.0
oauth2client==4.1.3
paypalrestsdk==1.13.1
Pillow==10.0.0
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.6.0
pyOpenSSL==23.1.0
pyparsing==3.1.0
Pyrebase4==4.7.1
python-jwt==4.0.0
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
wrapt==1.15.0
